fix(curation): GET collection_id returns the processing_status (#3026)
* fix(curation): GET collection_id returns the processing_status

- add function `add_collection_level_processing_status` to move the dataset processing status up to the collection level. This was removed from `list_collections_curation` to make it reusable.
- Move `add_collection_level_processing_status` and `list_collections_curation`  out of the Collection class. The entity class and all of its hellspawn are a horrible plague on our code base that should have never been created.
