+++ tests/unit/backend/corpora/api_server/curator/collection/test_curation_collections.py
@@ -283,6 +283,7 @@ class TestGetCollectionID(BaseAuthAPITest):
             },
         ],
         "name": "test_collection_name",
+        "processing_status": "PENDING",
         "published_at": None,
         "publisher_metadata": None,
         "revised_at": None,
